refactor: colocate quest instance service tests
